+++ packages/core/CHANGELOG.md
@@ -3,6 +3,99 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.16.3](https://github.com/IBM/carbon-charts/compare/v0.16.2...v0.16.3) (2019-09-11)
+
+
+### Bug Fixes
+
+* **angular:** Fix angular SSR issues ([86aa9d3](https://github.com/IBM/carbon-charts/commit/86aa9d3))
+
+
+
+
+
+## [0.16.2](https://github.com/IBM/carbon-charts/compare/v0.16.1...v0.16.2) (2019-09-09)
+
+
+### Bug Fixes
+
+* **core, react:** fix react SSR & legend sizing issues ([abedba0](https://github.com/IBM/carbon-charts/commit/abedba0))
+
+
+
+
+
+## [0.16.1](https://github.com/IBM/carbon-charts/compare/v0.16.0...v0.16.1) (2019-09-06)
+
+**Note:** Version bump only for package @carbon/charts
+
+
+
+
+
+# [0.16.0](https://github.com/IBM/carbon-charts/compare/v0.15.8...v0.16.0) (2019-09-06)
+
+
+### Features
+
+* **core:** Expand slice arcs in pie and donut on hover, and remove strokes ([ca5a467](https://github.com/IBM/carbon-charts/commit/ca5a467))
+* **core:** support width & maxWidth values for bars ([a2e7120](https://github.com/IBM/carbon-charts/commit/a2e7120))
+
+
+
+
+
+## [0.15.8](https://github.com/IBM/carbon-charts/compare/v0.15.7...v0.15.8) (2019-08-24)
+
+**Note:** Version bump only for package @carbon/charts
+
+
+
+
+
+## [0.15.7](https://github.com/IBM/carbon-charts/compare/v0.15.6...v0.15.7) (2019-08-22)
+
+**Note:** Version bump only for package @carbon/charts
+
+
+
+
+
+## [0.15.6](https://github.com/IBM/carbon-charts/compare/v0.15.5...v0.15.6) (2019-08-19)
+
+**Note:** Version bump only for package @carbon/charts
+
+
+
+
+
+## [0.15.5](https://github.com/IBM/carbon-charts/compare/v0.15.4...v0.15.5) (2019-08-19)
+
+**Note:** Version bump only for package @carbon/charts
+
+
+
+
+
+## [0.15.4](https://github.com/IBM/carbon-charts/compare/v0.15.3...v0.15.4) (2019-08-16)
+
+
+### Bug Fixes
+
+* **core:** fix resize grid backdrop ([8e1211e](https://github.com/IBM/carbon-charts/commit/8e1211e))
+
+
+
+
+
+## [0.15.3](https://github.com/IBM/carbon-charts/compare/v0.15.2...v0.15.3) (2019-08-16)
+
+**Note:** Version bump only for package @carbon/charts
+
+
+
+
+
 ## [0.15.2](https://github.com/IBM/carbon-charts/compare/v0.15.1...v0.15.2) (2019-08-12)